fix coins table and coin model

+++ database/migrations/2022_02_22_075413_add_columns_to_coins_table.php
@@ -17,7 +17,7 @@ class AddColumnsToCoinsTable extends Migration
             $table->string('name');
             $table->string('full_name');
             $table->text('icon_url');
-            $table->unsignedBigInteger('price_usd');
+            $table->float('price_usd');
             $table->text('additional_data');
         });
     }
